Renamed stages.py to files.py
