+++ src/lib/systemManagers/timeManager.py
@@ -1,6 +1,6 @@
 import json
 from datetime import datetime
-from lib.processing.stages import Step
+from lib.processing.files import Step
 import json
 from pathlib import Path
