+++ src/tools/convert.py
@@ -1,5 +1,5 @@
 from lib.data.argParser import ArgParser
-from lib.processing.stages import Step
+from lib.processing.files import Step
 
 if __name__ == '__main__':
     parser = ArgParser(
